# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing Cloud Hypervisor VMs.",
        "title": "Cloud Hypervisor API",
        "termsOfService": "https://github.com/org/ch-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/org/ch-api/issues",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vms/{id}/boot": {
            "post": {
                "description": "Boots the virtual machine with the given ID.",
                "tags": [
                    "VM Lifecycle"
                ],
                "summary": "Boot VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
...
        "/api/v1/vms/{id}/pause": {
            "post": {
                "description": "Pauses the virtual machine with the given ID.",
                "tags": [
                    "VM Lifecycle"
                ],
                "summary": "Pause VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{id}/reboot": {
            "post": {
                "description": "Reboots the virtual machine with the given ID.",
                "tags": [
                    "VM Lifecycle"
                ],
                "summary": "Reboot VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{id}/resume": {
            "post": {
                "description": "Resumes the paused virtual machine with the given ID.",
                "tags": [
                    "VM Lifecycle"
                ],
                "summary": "Resume VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{id}/shutdown": {
            "post": {
                "description": "Shuts down the virtual machine with the given ID.",
                "tags": [
                    "VM Lifecycle"
                ],
                "summary": "Shutdown VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "problem.Detail": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence.",
                    "type": "string"
                },
                "errors": {
                    "description": "Errors holds machine-readable validation errors.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/problem.FieldError"
                    }
                },
                "instance": {
                    "description": "Instance is a URI reference that identifies the specific occurrence.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code generated by the origin server.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI reference that identifies the problem type.",
                    "type": "string"
                }
            }
        },
        "problem.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
...
}